requirements.txt
awscli==1.10.6
twython==3.3.0
django-debug-toolbar==1.4
onespacemedia-cms==1.12.4
pathlib==1.0.1
praw==3.2.1
django-braces==1.8.1
django-filter==0.10.0
django-rest-swagger==0.3.3
django-s3-storage==0.9.8
django>=1.8,<1.9
git+https://github.com/danielsamuels/pyrope.git
django-suit==0.2.18
django-compressor==2.0
django-reversion==1.8.7
django-historylinks
django-watson==1.1.9
django-extensions
django-cachalot
python-social-auth
Werkzeug
